requirement.txt
transformers
torch
streamlit
pandas
scikit-learn
nltk
matplotlib
wordcloud
google-generativeai